Logic fixed, main.cpp added, Cmake complete.
