+++ Boardfiles/nucleol552zeq/.settings/stm32cubeide.project.prefs
@@ -1,5 +1,5 @@
 635E684B79701B039C64EA45C3F84D30=133CEB6F04A68B38073C17D67206B6A5
 66BE74F758C12D739921AEA421D593D3=0
 8DF89ED150041C4CBC7CB9A9CAA90856=26D62537746E9FFBD9A8CC3638B113BC
-DC22A860405A8BF2F2C095E5B6529F12=E03D1377FF9DFAFFEE5681C9DDDBDE4E
+DC22A860405A8BF2F2C095E5B6529F12=26D62537746E9FFBD9A8CC3638B113BC
 eclipse.preferences.version=1
