+++ Boardfiles/nucleol552zeq/Core/Inc/stm32l5xx_it.h
@@ -57,6 +57,8 @@ void DMA1_Channel2_IRQHandler(void);
 void DMA1_Channel3_IRQHandler(void);
 void TIM5_IRQHandler(void);
 void USART3_IRQHandler(void);
+void DMA1_Channel4_IRQHandler(void);
+void USART2_IRQHandler(void);
 /* USER CODE BEGIN EFP */
 
 /* USER CODE END EFP */
