+++ Boardfiles/nucleol552zeq/nucleol552zeq.ioc
@@ -14,6 +14,9 @@ Dma.Request0=TIM5_CH1
 Dma.Request1=TIM5_CH2
 Dma.Request2=USART3_RX
 Dma.RequestsNb=3
+Dma.Request2=USART2_RX
+Dma.Request3=USART2_TX
+Dma.RequestsNb=4
 Dma.TIM5_CH1.0.Channel_PRIV_NPRIV=DMA_CHANNEL_NPRIV_DISABLE
 Dma.TIM5_CH1.0.Direction=DMA_MEMORY_TO_PERIPH
 Dma.TIM5_CH1.0.EventEnable=DISABLE
@@ -50,24 +53,42 @@ Dma.TIM5_CH2.1.SyncEnable=DISABLE
 Dma.TIM5_CH2.1.SyncPolarity=HAL_DMAMUX_SYNC_NO_EVENT
 Dma.TIM5_CH2.1.SyncRequestNumber=1
 Dma.TIM5_CH2.1.SyncSignalID=NONE
-Dma.USART3_RX.2.Channel_PRIV_NPRIV=DMA_CHANNEL_NPRIV_DISABLE
-Dma.USART3_RX.2.Direction=DMA_PERIPH_TO_MEMORY
-Dma.USART3_RX.2.EventEnable=DISABLE
-Dma.USART3_RX.2.Instance=DMA1_Channel3
-Dma.USART3_RX.2.MemDataAlignment=DMA_MDATAALIGN_BYTE
-Dma.USART3_RX.2.MemInc=DMA_MINC_ENABLE
-Dma.USART3_RX.2.Mode=DMA_NORMAL
-Dma.USART3_RX.2.PeriphDataAlignment=DMA_PDATAALIGN_BYTE
-Dma.USART3_RX.2.PeriphInc=DMA_PINC_DISABLE
-Dma.USART3_RX.2.Polarity=HAL_DMAMUX_REQ_GEN_RISING
-Dma.USART3_RX.2.Priority=DMA_PRIORITY_LOW
-Dma.USART3_RX.2.RequestNumber=1
-Dma.USART3_RX.2.RequestParameters=Instance,Direction,PeriphInc,MemInc,PeriphDataAlignment,MemDataAlignment,Mode,Priority,SignalID,Polarity,RequestNumber,SyncSignalID,SyncPolarity,SyncEnable,EventEnable,SyncRequestNumber,Channel_PRIV_NPRIV
-Dma.USART3_RX.2.SignalID=NONE
-Dma.USART3_RX.2.SyncEnable=DISABLE
-Dma.USART3_RX.2.SyncPolarity=HAL_DMAMUX_SYNC_NO_EVENT
-Dma.USART3_RX.2.SyncRequestNumber=1
-Dma.USART3_RX.2.SyncSignalID=NONE
+Dma.USART2_RX.2.Channel_PRIV_NPRIV=DMA_CHANNEL_NPRIV_DISABLE
+Dma.USART2_RX.2.Direction=DMA_PERIPH_TO_MEMORY
+Dma.USART2_RX.2.EventEnable=DISABLE
+Dma.USART2_RX.2.Instance=DMA1_Channel3
+Dma.USART2_RX.2.MemDataAlignment=DMA_MDATAALIGN_BYTE
+Dma.USART2_RX.2.MemInc=DMA_MINC_ENABLE
+Dma.USART2_RX.2.Mode=DMA_CIRCULAR
+Dma.USART2_RX.2.PeriphDataAlignment=DMA_PDATAALIGN_BYTE
+Dma.USART2_RX.2.PeriphInc=DMA_PINC_DISABLE
+Dma.USART2_RX.2.Polarity=HAL_DMAMUX_REQ_GEN_RISING
+Dma.USART2_RX.2.Priority=DMA_PRIORITY_LOW
+Dma.USART2_RX.2.RequestNumber=1
+Dma.USART2_RX.2.RequestParameters=Instance,Direction,PeriphInc,MemInc,PeriphDataAlignment,MemDataAlignment,Mode,Priority,SignalID,Polarity,RequestNumber,SyncSignalID,SyncPolarity,SyncEnable,EventEnable,SyncRequestNumber,Channel_PRIV_NPRIV
+Dma.USART2_RX.2.SignalID=NONE
+Dma.USART2_RX.2.SyncEnable=DISABLE
+Dma.USART2_RX.2.SyncPolarity=HAL_DMAMUX_SYNC_NO_EVENT
+Dma.USART2_RX.2.SyncRequestNumber=1
+Dma.USART2_RX.2.SyncSignalID=NONE
+Dma.USART2_TX.3.Channel_PRIV_NPRIV=DMA_CHANNEL_NPRIV_DISABLE
+Dma.USART2_TX.3.Direction=DMA_MEMORY_TO_PERIPH
+Dma.USART2_TX.3.EventEnable=DISABLE
+Dma.USART2_TX.3.Instance=DMA1_Channel4
+Dma.USART2_TX.3.MemDataAlignment=DMA_MDATAALIGN_BYTE
+Dma.USART2_TX.3.MemInc=DMA_MINC_ENABLE
+Dma.USART2_TX.3.Mode=DMA_CIRCULAR
+Dma.USART2_TX.3.PeriphDataAlignment=DMA_PDATAALIGN_BYTE
+Dma.USART2_TX.3.PeriphInc=DMA_PINC_DISABLE
+Dma.USART2_TX.3.Polarity=HAL_DMAMUX_REQ_GEN_RISING
+Dma.USART2_TX.3.Priority=DMA_PRIORITY_LOW
+Dma.USART2_TX.3.RequestNumber=1
+Dma.USART2_TX.3.RequestParameters=Instance,Direction,PeriphInc,MemInc,PeriphDataAlignment,MemDataAlignment,Mode,Priority,SignalID,Polarity,RequestNumber,SyncSignalID,SyncPolarity,SyncEnable,EventEnable,SyncRequestNumber,Channel_PRIV_NPRIV
+Dma.USART2_TX.3.SignalID=NONE
+Dma.USART2_TX.3.SyncEnable=DISABLE
+Dma.USART2_TX.3.SyncPolarity=HAL_DMAMUX_SYNC_NO_EVENT
+Dma.USART2_TX.3.SyncRequestNumber=1
+Dma.USART2_TX.3.SyncSignalID=NONE
 FREERTOS.IPParameters=Tasks01
 FREERTOS.Tasks01=defaultTask,24,128,StartDefaultTask,Default,NULL,Dynamic,NULL,NULL
 File.Version=6
@@ -181,6 +202,7 @@ NVIC.BusFault_IRQn=true\:0\:0\:false\:false\:true\:false\:false\:false\:false
 NVIC.DMA1_Channel1_IRQn=true\:5\:0\:false\:false\:true\:true\:false\:true\:true
 NVIC.DMA1_Channel2_IRQn=true\:5\:0\:false\:false\:true\:true\:false\:true\:true
 NVIC.DMA1_Channel3_IRQn=true\:5\:0\:false\:false\:true\:true\:false\:true\:true
+NVIC.DMA1_Channel4_IRQn=true\:5\:0\:false\:false\:true\:true\:false\:true\:true
 NVIC.DebugMonitor_IRQn=true\:0\:0\:false\:false\:true\:false\:false\:false\:false
 NVIC.ForceEnableDMAVector=true
 NVIC.HardFault_IRQn=true\:0\:0\:false\:false\:true\:false\:false\:false\:false
@@ -195,6 +217,7 @@ NVIC.SavedSystickIrqHandlerGenerated=true
 NVIC.SysTick_IRQn=true\:7\:0\:false\:false\:false\:true\:false\:true\:false
 NVIC.TIM5_IRQn=true\:5\:0\:false\:false\:true\:true\:true\:true\:true
 NVIC.USART3_IRQn=true\:5\:0\:false\:false\:true\:true\:true\:true\:true
+NVIC.USART2_IRQn=true\:5\:0\:false\:false\:true\:true\:true\:true\:true
 NVIC.UsageFault_IRQn=true\:0\:0\:false\:false\:true\:false\:false\:false\:false
 PA0.Signal=S_TIM2_CH1
 PA1.Signal=S_TIM2_CH2
@@ -342,7 +365,7 @@ ProjectManager.ToolChainLocation=
 ProjectManager.UAScriptAfterPath=
 ProjectManager.UAScriptBeforePath=
 ProjectManager.UnderRoot=true
-ProjectManager.functionlistsort=1-SystemClock_Config-RCC-false-HAL-false,2-MX_GPIO_Init-GPIO-false-HAL-true,3-MX_DMA_Init-DMA-false-HAL-true,4-MX_LPUART1_UART_Init-LPUART1-false-HAL-true,5-MX_RTC_Init-RTC-false-HAL-true,6-MX_UCPD1_Init-UCPD1-false-LL-true,7-MX_USB_PCD_Init-USB-false-HAL-true,8-MX_I2C1_Init-I2C1-false-HAL-true,9-MX_I2C2_Init-I2C2-false-HAL-true,10-MX_UART5_Init-UART5-false-HAL-true,11-MX_USART1_UART_Init-USART1-false-HAL-true,12-MX_USART2_UART_Init-USART2-false-HAL-true,13-MX_USART3_UART_Init-USART3-false-HAL-true,14-MX_SPI1_Init-SPI1-false-HAL-true,15-MX_SPI2_Init-SPI2-false-HAL-true,16-MX_TIM1_Init-TIM1-false-HAL-true,17-MX_TIM2_Init-TIM2-false-HAL-true,18-MX_TIM4_Init-TIM4-false-HAL-true,19-MX_TIM5_Init-TIM5-false-HAL-true,20-MX_ADC1_Init-ADC1-false-HAL-true,21-MX_ICACHE_Init-ICACHE-false-HAL-true,22-MX_IWDG_Init-IWDG-false-HAL-true,0-MX_PWR_Init-PWR-false-HAL-true
+ProjectManager.functionlistsort=1-SystemClock_Config-RCC-false-HAL-false,2-MX_GPIO_Init-GPIO-false-HAL-true,3-MX_DMA_Init-DMA-false-HAL-true,4-MX_LPUART1_UART_Init-LPUART1-false-HAL-true,5-MX_RTC_Init-RTC-false-HAL-true,6-MX_UCPD1_Init-UCPD1-false-LL-true,7-MX_USB_PCD_Init-USB-false-HAL-true,8-MX_I2C1_Init-I2C1-false-HAL-true,9-MX_I2C2_Init-I2C2-false-HAL-true,10-MX_UART5_Init-UART5-false-HAL-true,11-MX_USART1_UART_Init-USART1-false-HAL-true,12-MX_USART2_UART_Init-USART2-false-HAL-true,13-MX_USART3_UART_Init-USART3-false-HAL-true,14-MX_SPI1_Init-SPI1-false-HAL-true,15-MX_SPI2_Init-SPI2-false-HAL-true,16-MX_TIM1_Init-TIM1-false-HAL-true,17-MX_TIM2_Init-TIM2-false-HAL-true,18-MX_TIM4_Init-TIM4-false-HAL-true,19-MX_TIM5_Init-TIM5-false-HAL-true,20-MX_ADC1_Init-ADC1-false-HAL-true,21-MX_ICACHE_Init-ICACHE-false-HAL-true,22-MX_IWDG_Init-IWDG-false-HAL-true,23-MX_TIM3_Init-TIM3-false-HAL-true,0-MX_PWR_Init-PWR-false-HAL-true
 RCC.ADCCLockSelection=RCC_ADCCLKSOURCE_SYSCLK
 RCC.ADCFreq_Value=110000000
 RCC.AHBFreq_Value=110000000
