+++ Tools/Firmware/CMakeLists.txt
@@ -87,6 +87,7 @@ set(DRIVERS_Common_CXX_SOURCES "${DRIVERS_Common_DIR}/*.cpp")
 set(DRIVERS_IWDGDriver_DIR ${ROOT_DIR}/Drivers/IWDG_Driver/Src)
 set(DRIVERS_IWDGDriver_CXX_SOURCES "${DRIVERS_IWDGDriver_DIR}/*.cpp")
 
+
 ## Actually find the sources, NOTE: if you add a new source above, add it here! ##
 ## Drivers Sources ##
 set(DRIVERS_UARTDevice_DIR ${ROOT_DIR}/Drivers/UARTDevice/Src)
